dev remove unused functions and files
