--- src/services/userService.js
@@ -1,58 +1,0 @@
-// import Boom from 'boom';
-// import User from '../models/user';
-
-/**
- * Get all users.
- *
- * @return {Promise}
- */
-// export function getAllUsers() {
-//   return User.fetchAll();
-// }
-
-// /**
-//  * Get a user.
-//  *
-//  * @param  {Number|String}  id
-//  * @return {Promise}
-//  */
-// export function getUser(id) {
-//   return new User({ id }).fetch().then(user => {
-//     if (!user) {
-//       throw Boom.notFound('User not found');
-//     }
-
-//     return user;
-//   });
-// }
-
-// /**
-//  * Create new user.
-//  *
-//  * @param  {Object}  user
-//  * @return {Promise}
-//  */
-// export function createUser(user) {
-//   return new User({ name: user.name }).save();
-// }
-
-// /**
-//  * Update a user.
-//  *
-//  * @param  {Number|String}  id
-//  * @param  {Object}         user
-//  * @return {Promise}
-//  */
-// export function updateUser(id, user) {
-//   return new User({ id }).save({ name: user.name });
-// }
-
-// /**
-//  * Delete a user.
-//  *
-//  * @param  {Number|String}  id
-//  * @return {Promise}
-//  */
-// export function deleteUser(id) {
-//   return new User({ id }).fetch().then(user => user.destroy());
-// }
